pyproject.toml
[project]
name = "velogs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "matplotlib>=3.10.8",
    "pyserial>=3.5",
]

[project.scripts]
visualgs = "ui.cli2:main"

[tool.setuptools]
packages = ["ui", "telemetry", "health"]